backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Antoine Debroye <ant.debroye@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
flask-cors = "^5.0.0"
flask-bcrypt = "^1.0.1"
pyjwt = "^2.9.0"
flask-jwt-extended = "^4.6.0"
flask-socketio = "^5.4.1"
eventlet = "^0.37.0"
pyyaml = "^6.0.2"
pylint = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
